fix: cherry-pick from cactbot e897b780aff7f53c9f723c7407c824516a6e095a

+++ OverlayPlugin.Core/NetworkProcessors/FateWatcher.cs
@@ -115,7 +115,7 @@ namespace RainbowMage.OverlayPlugin.NetworkProcessors
 
         public unsafe void ProcessMessage(byte* buffer, byte[] message)
         {
-            int a = *((int*)&buffer[Category_Offset]);
+            int a = *(ushort*)&buffer[Category_Offset];
 
             if (a == opcodes[region_].add)
             {
